isl_schedule_constraints: add prefix schedule constraint
The prefix schedule constraint allows the scheduler to take
into account some outer predetermined part of the schedule.
The generated schedule will only schedule those instances
apart that are not already scheduled apart by the prefix schedule.
For example, it could be used to reschedule a subtree of a schedule tree
by taking into account the prefix schedule at the corresponding node.

The prefix schedule is represented as an isl_multi_union_pw_aff
because this most closely matches how a prefix schedule is represented
in a schedule tree.  This choice also allows the absence of a prefix
schedule to be represented in a way that is independent
of the schedule domain, reducing the overhead for the case
where no prefix schedule is specified.
A disadvantage is that it can more easily lead to
malformed inputs, if they are not derived directly from a schedule tree.
Alternative representations would have been an isl_union_pw_multi_aff
or an isl_union_map.

For the purpose of avoid linealy dependent schedule rows,
the linear part of the prefix schedule is stored as
an initial band of the per-node schedule.
The effect is similar to that of a trivial carrying band.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>